[CRT-420] Use unified student reference hook
Remove source-specific branching when starring analyses in the frontend.
