--- frontend/src/api/collimator/hooks/solutions/usePatchStudentActivityIsReference.ts
@@ -1,66 +1,0 @@
-import { useCallback } from "react";
-import { useSWRConfig } from "swr";
-import {
-  getSolutionsControllerFindCurrentAnalysesV0Url,
-  solutionsControllerPatchStudentActivityIsReferenceV0,
-} from "../../generated/endpoints/solutions/solutions";
-import { useAuthenticationOptions } from "../authentication/useAuthenticationOptions";
-import { CurrentStudentAnalysis } from "../../models/solutions/current-student-analysis";
-import { GetCurrentAnalysisReturnType } from "./useCurrentSessionTaskSolutions";
-import { useRevalidateSolutionList } from "./useRevalidateSolutionList";
-
-type PatchActivityType = (
-  classId: number,
-  sessionId: number,
-  taskId: number,
-  studentId: number,
-  solutionHash: string,
-  isReference: boolean,
-) => Promise<void>;
-
-export const usePatchStudentActivityIsReference = (): PatchActivityType => {
-  const revalidateSolutionList = useRevalidateSolutionList();
-  const { mutate, cache } = useSWRConfig();
-  const authOptions = useAuthenticationOptions();
-
-  return useCallback<PatchActivityType>(
-    (classId, sessionId, taskId, studentId, solutionHash, isReference) =>
-      solutionsControllerPatchStudentActivityIsReferenceV0(
-        classId,
-        sessionId,
-        taskId,
-        studentId,
-        { isReference, solutionHash },
-        undefined,
-        authOptions,
-      ).then(() => {
-        revalidateSolutionList(classId, sessionId, taskId);
-
-        // Optimistically update the analysis list cache
-        const key = getSolutionsControllerFindCurrentAnalysesV0Url(
-          classId,
-          sessionId,
-          taskId,
-          {},
-        );
-
-        const cachedData: GetCurrentAnalysisReturnType | undefined =
-          cache.get(key)?.data;
-
-        if (cachedData !== undefined) {
-          mutate(
-            key,
-            cachedData.map((a) =>
-              a instanceof CurrentStudentAnalysis &&
-              !a.isStudentSolution &&
-              a.studentId === studentId &&
-              a.solutionHash === solutionHash
-                ? a.withIsReference(isReference)
-                : a,
-            ),
-          );
-        }
-      }),
-    [authOptions, revalidateSolutionList, cache, mutate],
-  );
-};
