+++ frontend/src/api/collimator/hooks/solutions/useStarAnalysis.ts
@@ -1,7 +1,6 @@
 import { useCallback } from "react";
 import { CurrentStudentAnalysis } from "../../models/solutions/current-student-analysis";
-import { usePatchStudentSolutionIsReference } from "./usePatchStudentSolutionIsReference";
-import { usePatchStudentActivityIsReference } from "./usePatchStudentActivityIsReference";
+import { usePatchStudentReferenceSolution } from "./usePatchStudentReferenceSolution";
 
 type StarAnalysisType = (
   classId: number,
@@ -10,30 +9,18 @@ type StarAnalysisType = (
 ) => Promise<void>;
 
 export const useStarAnalysis = (): StarAnalysisType => {
-  const patchStudentSolutionIsReference = usePatchStudentSolutionIsReference();
-  const patchStudentActivityIsReference = usePatchStudentActivityIsReference();
+  const patchStudentReferenceSolution = usePatchStudentReferenceSolution();
 
   return useCallback<StarAnalysisType>(
-    (classId, analysis, isReference) => {
-      if (analysis.isStudentSolution) {
-        return patchStudentSolutionIsReference(
-          classId,
-          analysis.sessionId,
-          analysis.taskId,
-          analysis.studentSolutionId!,
-          { isReference },
-        );
-      }
-
-      return patchStudentActivityIsReference(
+    (classId, analysis, isReference) =>
+      patchStudentReferenceSolution(
         classId,
         analysis.sessionId,
         analysis.taskId,
         analysis.studentId,
         analysis.solutionHash,
         isReference,
-      );
-    },
-    [patchStudentSolutionIsReference, patchStudentActivityIsReference],
+      ),
+    [patchStudentReferenceSolution],
   );
 };
